app changed so that it will connect to the api

+++ RatApplication/app/src/main/java/cs2340/ratapplication/models/Sighting.java
@@ -16,7 +16,7 @@ public class Sighting {
     public String city;
     public String borough;
     public int zip;
-    public float latitude;
-    public float longitude;
+    public double latitude;
+    public double longitude;
 
 }
